Controller remake
