+++ .gitignore
@@ -1,4 +1,5 @@
 # Common IntelliJ Platform excludes
+*Development.json
 
 # User specific
 **/.idea/**/workspace.xml
