+++ src/CommunicationService/Classifications/Api/ClassificationGetByIdController.cs
@@ -1,0 +1,25 @@
+using CommunicationService.Classifications.Fundamental;
+
+namespace CommunicationService.Classifications.Api;
+
+[ApiController]
+[Route("[controller]")]
+public class ClassificationGetByIdController : ClassificationBaseController
+{
+    private IClassificationRepository ClassificationRepository { get; }
+
+    public ClassificationGetByIdController(IClassificationRepository classificationRepository)
+    {
+        ClassificationRepository = classificationRepository;
+    }
+
+    [HttpGet("ById/{id:guid}")]
+    public async Task<IActionResult> GetClassificationById(Guid id, CancellationToken cancellationToken)
+    {
+        var classificationResult = await ClassificationRepository.GetClassificationById(id, cancellationToken);
+
+        return classificationResult.Match(
+            item => Ok(item.ToClassificationResponse()),
+            Problem);
+    }
+}
