+++ src/CommunicationService/Classifications/Api/Models/ClassificationResponse.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.Classifications.ContractModels;
+namespace CommunicationService.Classifications.Api.Models;
 
 public record ClassificationResponse (Guid Id, string Name, string[] MetadataTypes);
