+++ src/CommunicationService/Classifications/Api/Models/CreateClassificationRequest.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.Classifications.ContractModels;
+namespace CommunicationService.Classifications.Api.Models;
 
 public record CreateClassificationRequest(string Name, string[] MetadataTypes);
