+++ src/CommunicationService/Classifications/Api/Models/UpsertClassificationRequest.cs
@@ -1,4 +1,4 @@
-namespace CommunicationService.Classifications.ContractModels;
+namespace CommunicationService.Classifications.Api.Models;
 
 public record UpsertClassificationRequest(string Name, string[] MetadataTypes)
 {
