+++ src/CommunicationService/Classifications/Fundamental/ApiRouteHelper.cs
@@ -1,0 +1,6 @@
+namespace CommunicationService.Classifications.Fundamental;
+
+public static class ApiRouteHelper
+{
+    public const string ClassificationGetByIdRoute = "ClassificationGetById";
+}
