+++ src/CommunicationService/MetadataTypes/Api/DataContract/CreateMetadataTypeRequest.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.MetadataTypes.Contracts;
+namespace CommunicationService.MetadataTypes.Api.DataContract;
 
 public record CreateMetadataTypeRequest(string Name, string[] Classifications);
