+++ src/CommunicationService/MetadataTypes/Api/DataContract/MetadataTypeResponse.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.MetadataTypes.Contracts;
+namespace CommunicationService.MetadataTypes.Api.DataContract;
 
 public record MetadataTypeResponse (Guid Id, string Name, string[] Classifications);
