+++ src/CommunicationService/MetadataTypes/Api/DataContract/UpsertClassificationRequest.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.MetadataTypes.Contracts;
+namespace CommunicationService.MetadataTypes.Api.DataContract;
 
 public record UpsertMetadataTypeRequest(string Name, string[] Classifications);
