+++ src/CommunicationService/MetadataTypes/Api/DataContract/UpsertedMetadataTypeResult.cs
@@ -1,3 +1,3 @@
-namespace CommunicationService.MetadataTypes.Models;
+namespace CommunicationService.MetadataTypes.Api.DataContract;
 
 public record struct UpsertedMetadataTypeResult(bool RegisteredAsNewItem);
