+++ src/CommunicationService/MetadataTypes/MetadataTypeConverter.cs
@@ -1,5 +1,5 @@
-using CommunicationService.MetadataTypes.Contracts;
-using CommunicationService.MetadataTypes.Models;
+using CommunicationService.MetadataTypes.Api.DataContract;
+using CommunicationService.MetadataTypes.DataModels;
 
 namespace CommunicationService.MetadataTypes;
