+++ tests/CommunicationService.Test/CommunicationService.Test.csproj
@@ -41,6 +41,10 @@
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         <CopyToPublishDirectory>Always</CopyToPublishDirectory>
       </Content>
+      <Content Update="appsettings.auto-test-developer.json">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        <CopyToPublishDirectory>Always</CopyToPublishDirectory>
+      </Content>
     </ItemGroup>
     
 </Project>
