+++ tests/CommunicationService.Test/MetadataTypeTests/Helpers/MetadataTypeFixture.cs
@@ -1,3 +1,5 @@
+using CommunicationService.MetadataTypes.DataModels;
+
 namespace CommunicationService.Test.MetadataTypeTests.Helpers;
 
 public class MetadataTypeFixture : BaseFixture
