+++ tests/CommunicationService.Test/ReceiversTests/Helpers/ReceiverDbContextHelper.cs
@@ -1,5 +1,6 @@
 using CommunicationService.Receivers.DataModels;
 using CommunicationService.Test.ClassificationTests;
+using CommunicationService.Test.ClassificationTests.Helpers;
 
 namespace CommunicationService.Test.ReceiversTests.Helpers;
